Ensure that wheezy has dotdeb repos
Since wheezy does not have php5.5 we will ensure that the dotdeb repos are
available to hosts running debian wheezy.  This will allow the this script to
successfully provision a host fully without any failed puppet runs. There was
discussion of whether this should happen post-provisioning, but if we did this
after this script ran puppet would be unable install the php dependencies until
the post-install script completed. This ensures that after the first run of the
bootstrap that a host will be fully provisioned.

ensure repo is added to apt